produce HDF5 files for edges defined in tsv files of triples (left, rel, right)

produce HDF5 files for edges defined in tsv files of triples (left, rel, right)

triples_to_hdf5_raw(
  triple_paths,
  num_partitions = 1L,
  relname = "all_edges",
  lhs = "all",
  rhs = "all",
  weight = 1,
  operator = "complex_diagonal",
  all_negs = FALSE,
  pbgref,
  pathlibref
)

Arguments

triple_paths

character(3) with test, train, valid edge sets in order

num_partitions

integer(1)

relname

character(1), passed to make_rel_schema

lhs

character(1), passed to make_rel_schema

rhs

character(1), passed to make_rel_schema

weight

numeric(1), passed to make_rel_schema

operator

character(1), passed to make_rel_schema

all_negs

character(1), passed to make_rel_schema

pbgref

python reference to torchbiggraph module

pathlibref

python reference to pathlib module

Examples

tfis = fb15k_folder()
pbg = reticulate::import("torchbiggraph")
palib = reticulate::import("pathlib")
nn = triples_to_hdf5(tfis, pbgref = pbg, pathlibref=palib)
#> Error in triples_to_hdf5(tfis, pbgref = pbg, pathlibref = palib): unused argument (pathlibref = palib)
nn
#> Error in eval(expr, envir, enclos): object 'nn' not found
rhdf5::h5ls(dir(nn$edge_paths[[1]], full=TRUE))
#> Error in eval(expr, envir, enclos): object 'nn' not found